Moved over to using content input for the Vectorizer, added pre-processing
Moved functionality over so that it reads files into a list before doing the vectorizing, meaning that it works with lists of strings instead of raw file objects. This makes it so a string can be used for a query instead of a file, which will be better for Slackbot queries.

Also added pre-processing on both the documents and the queries. Tried out lemmatization, stemming and removing stopwords - only lemmatization seemed to provide better results than no processing (removing stopwords made it _actively_ worse, stemming was more of a middling impact). WMay adjust when looking at all docs