Lots of changes in this one, I don't know the full list...
Moved everything over to draw from config instead of hard-coded stuff.

Tweaked how things get returned so we look through each configured directory individually to make it easier to determine which repo the docs came from. This might want revisiting in future, but was the easiest way I could think to do it right now

Tweaked what documents gets included in the list to be returned - we already take the top 5, but now if any of those have a similarity score of 0, we wipe them out and don't return them. If all 5 are empty, we just don't return anything and don't make the bot send a message. This will be extended in future to allow controlling what the bot replies to